% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{The application User-Interface}
\usage{
app_ui(request)
}
\arguments{
\item{request}{Internal parameter for \code{{shiny}}.}
}
\description{
The application user-interface separated as pages with the first page containing
the upload module.
}
\details{
\itemize{
\item \href{https://rstudio.github.io/bslib}{bslib R package} providing tools for
customizing Bootstrap themes.
\item \href{https://bootswatch.com/}{Bootswatch theme} Free themes for bootstrap.
\item \href{https://shiny.rstudio.com/articles/layout-guide.html}{Shiny layout}
\item \href{https://rstudio.github.io/shinydashboard/index.html}{shinydashboard}
\item \href{https://github.com/dreamRs/shinyWidgets}{shinywidgets}
\item \href{https://rstudio.github.io/sortable/}{sortable}
\item \href{https://github.com/daattali/colourpicker}{colourpicker}
}

tests
\itemize{
\item \href{https://testthat.r-lib.org/index.html}{testthat}
\item \href{https://testthat.r-lib.org/articles/snapshotting.html}{testthat snapshot tests}
\item \href{https://shiny.rstudio.com/reference/shiny/1.5.0/testServer.html}{testServer}
\item \href{https://rstudio.github.io/shinytest/}{shinytest}
}
}
