% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete-store.R
\name{deleteStore}
\alias{deleteStore}
\title{Delete local storage key}
\usage{
deleteStore(session, name)
}
\arguments{
  \item{session}{The session paramter from the
  \code{\link{shinyServer}} function.} \item{name}{The
  name for this setting. This is the name you referenced on
  updateStore}
}
\description{
Send the name of the field to delete
}

