% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-store.R
\name{initStore}
\alias{initStore}
\title{Initialize shinyStore in an application's UI}
\usage{
initStore(id, namespace, privateKey = NULL)
}
\arguments{
\item{id}{The identifier of the store (determines the name of the store on the
\code{input} reactive object in your \code{server.R} file).}

\item{namespace}{Not a \code{namespace} in the traditional R sense, but rather
the identifying prefix to use when storing any object in the browser's local
storage. This is the only means by which we can separate data stored for
different applications that are being hosted at the same (sub-)domain.}

\item{privateKey}{the private key to use to decrypt data. Must be provided
if you're going to have any encrypted fields.}
}
\description{
Though shinyStore doesn't have any user-visible elements, you still must 
initialize it in your application's \code{ui.R} file. You should execute this
function in your ui.R file to setup the input element associated with your store.
}
