% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_display_image.R
\name{mod_display_image_server}
\alias{mod_display_image_server}
\title{Shiny module display image server function}
\usage{
mod_display_image_server(id, mod_values)
}
\description{
This shiny module uses the filtered dataset and selected species from the filter
module to
\itemize{
\item display html link to the original image.
\item display an image of the penguin species.
}
}
\details{
\itemize{
\item \href{https://shiny.rstudio.com/articles/images.html}{Render images in a Shiny app}
}
}
