% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_display_image.R
\name{mod_display_image_ui}
\alias{mod_display_image_ui}
\title{display image UI Function}
\usage{
mod_display_image_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\description{
Shiny module to display html text and image using htmlOutput and imageOutput
controls.
}
\details{
\href{https://shiny.rstudio.com/articles/images.html}{Render images in a Shiny app}
}
