% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_display_plot.R
\name{mod_display_plot_ui}
\alias{mod_display_plot_ui}
\title{Shiny module display and download plot UI Function}
\usage{
mod_display_plot_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\description{
Shiny module to display and save a plot using plotOutput and downloadHandler controls
and table from brush option using tableOutput control.
}
\details{
Progress and spinner bars
\itemize{
\item \href{https://shiny.rstudio.com/articles/progress.html}{shiny withProgress}
\item \href{https://github.com/daattali/shinycssloaders}{shinycssLoaders}
\item \link{waiter}(waiter: https://github.com/JohnCoene/waiter)
}

Plots
\itemize{
\item \href{https://gallery.shinyapps.io/095-plot-interaction-advanced}{plot input}
\itemize{
\item click = clickOpts(id = "plot_click", ...)
\item dblclick = dblClickOpts(id = "plot_dblclick", ...)
\item hover = hoverOpts(id = "plot_hover", ...)
\item brush = brushOpts(id = "plot_brush", ...)
}
}
}
