% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_display.R
\name{mod_display_server}
\alias{mod_display_server}
\title{Shiny module display server function}
\usage{
mod_display_server(id, mod_values)
}
\description{
This shiny module uses the filtered dataset and selected species from the filter
module to display an image of the penguin species with a html link to the original
image.
}
\details{
\itemize{
\item \href{https://shiny.rstudio.com/articles/images.html}{Render images in a Shiny app}
}
}
