% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_display_table.R
\name{mod_display_table_server}
\alias{mod_display_table_server}
\title{Shiny module display table server function}
\usage{
mod_display_table_server(id, mod_values)
}
\description{
This shiny module displays a table of the filtered data,
\itemize{
\item check validated filter data available and has permission to display table.
\item displays the table of the filtered data.
}
}
\details{
\itemize{
\item \href{https://datatables.net/reference/option}{DataTables options}
}
}
