% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_display_table.R
\name{mod_display_table_ui}
\alias{mod_display_table_ui}
\title{display table UI Function}
\usage{
mod_display_table_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\description{
Shiny module to display and download table using dataTableOutput and downloadHandler
control.
}
\details{
\itemize{
\item \href{https://glin.github.io/reactable}{reactable tables}
}
}
