% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_display.R
\name{mod_display_ui}
\alias{mod_display_ui}
\title{display UI Function}
\usage{
mod_display_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\description{
Shiny module to display html text and image using htmlOutput and imageOutput
controls.
}
