% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter_permission.R
\name{mod_filter_permission_server}
\alias{mod_filter_permission_server}
\title{Shiny module filter permission server function}
\usage{
mod_filter_permission_server(id, mod_values, permission_required = TRUE)
}
\arguments{
\item{id}{Internal parameter for {shiny}.}

\item{mod_values}{reactiveValues() list containing server module returned values}
}
\value{
species_list(), year_list and filter_data() reactive value containing uploaded data from csv file.
}
\description{
This shiny module uses the dataset returned from the upload module to
\itemize{
\item populate species and year drop down options in the input controls.
\item filter the dataset by the selected species and year in the drop down option.
\item validate the input control options and filtered dataset.
\item dynamically add buttons seeking permission to display plot and table of filtered data.
(default, permission_required = TRUE), else skips this permission step (permission_required = FALSE),
\item send notification of returned reactive values.
}
}
\details{
\itemize{
\item \href{https://shiny.rstudio.com/reference/shiny/0.14/freezeReactiveValue.html}{freezeReactiveValue} prevents flicker in dynamic controls
\item \href{https://rstudio.github.io/shinyvalidate}{shinyValidate}
\item \href{https://github.com/merlinoa/shinyFeedback}{shinyFeedback}
\item \href{https://github.com/daattali/shinyalert}{shinyAlert}
}
}
