% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter_permission.R
\name{mod_filter_permission_ui}
\alias{mod_filter_permission_ui}
\title{Shiny module filter permission UI Function}
\usage{
mod_filter_permission_ui(id, dataset_name)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\description{
Shiny module to filter dataset using selectInput and numericInput controls and
manage permissions to display data.
}
\details{
\itemize{
\item \href{https://shiny.rstudio.com/articles/selectize.html#server-side-selectize}{Dynamic selectInput}
\item \href{https://shiny.rstudio.com/articles/sliders.html}{Using sliders}
}
}
