% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter.R
\name{mod_filter_server}
\alias{mod_filter_server}
\title{Shiny module filter server function}
\usage{
mod_filter_server(id, mod_values)
}
\arguments{
\item{id}{Internal parameter for {shiny}.}

\item{mod_values}{reactiveValues() list containing server module returned values}
}
\value{
species_list(), year_list and filter_data() reactive value containing uploaded data from csv file.
}
\description{
This shiny module uses the dataset returned from the upload module to populate
species and year drop down options in the input controls. After validating the
input values it returns the dataset filtered on selected species and year.
}
\details{
\itemize{
\item \href{https://shiny.rstudio.com/reference/shiny/0.14/freezeReactiveValue.html}{freezeReactiveValue} prevents flicker in dynamic controls
\item \href{https://rstudio.github.io/shinyvalidate}{shinyValidate}
\item \href{https://github.com/merlinoa/shinyFeedback}{shinyFeedback}
\item \href{https://github.com/daattali/shinyalert}{shinyAlert}
}
}
