% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tour_guide.R
\name{mod_tour_guide_server}
\alias{mod_tour_guide_server}
\title{Shiny module tour guide server function}
\usage{
mod_tour_guide_server(id, mod_values)
}
\arguments{
\item{id}{Internal parameter for {shiny}.}

\item{mod_values}{reactiveValues() list containing server module returned values}
}
\description{
This shiny modules provides the tour guide information boxes using cicerone
package.
}
\details{
\itemize{
\item \href{https://cicerone.john-coene.com/}{cicerone} - creates guided tours using driver.js library.
\item \href{https://conductor.etiennebacher.com/}{conductor} - creates guided tours using shephard.js library.
\item \href{https://carlganz.github.io/rintrojs/}{rintrojs} - creates guided tours using Intro.js library.
\item \href{https://github.com/ColinFay/hexmake/blob/master/R/mod_guided_tour.R}{Colin Fay hexmake tour guide module}
}
}
