% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tour_guide.R
\name{mod_tour_guide_ui}
\alias{mod_tour_guide_ui}
\title{Shiny module tour guide function}
\usage{
mod_tour_guide_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module to display a walk through tour-guide of shiny modules
and their controls
}
