% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload.R
\name{mod_upload_ui}
\alias{mod_upload_ui}
\title{upload UI Function}
\usage{
mod_upload_ui(id)
}
\arguments{
\item{id}{Internal parameter for {shiny}.}
}
\description{
Shiny module uploading csv file using fileInput control.
}
\details{
\itemize{
\item \href{https://shiny.rstudio.com/articles/upload.html}{Uploading files} using fileInput control.
}
}
