% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{notify}
\alias{notify}
\title{Show notification}
\usage{
notify(msg, id = NULL, duration = NULL)
}
\arguments{
\item{msg}{notification text message to be displayed}

\item{id}{internal notification id}

\item{duration}{duration of notification message in seconds.}
}
\description{
Function showing a notification message on the bottom right of the User Interface
}
\details{
\itemize{
\item \href{https://shiny.rstudio.com/articles/notifications.html}{Shiiny notifications}
}
}
