% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\docType{data}
\name{species_images}
\alias{species_images}
\title{Penguin species images}
\format{
A tibble with 3 rows and 3 variables.
\describe{
\item{species}{Names of the penguin species.}
\item{id}{Photograph id taken used in the Unsplash website.}
\item{author}{Name of photographer used in the Unsplash website.}
}
}
\usage{
species_images
}
\description{
Images of the three penguin species in the \href{https://allisonhorst.github.io/palmerpenguins/}{palmerpenguin package}
dataset used in this shiny app example. The images are taken from \href{https://unsplash.com/}{Unsplash},
an outline source of royalty-free stock photographs.
}
\keyword{datasets}
