% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ssErr}
\alias{ssErr}
\title{Get shinyStore error code}
\usage{
ssErr(obj)
}
\arguments{
  \item{obj}{The object to inspect for an error.}
}
\description{
Given a shinyStore object, gets the error code off of it.
Will return 0 if no error, or a positive integer if there
is an error.
}

