% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\docType{data}
\name{tour_guide}
\alias{tour_guide}
\title{Shiny tour guide}
\format{
An object of class \code{Conductor} (inherits from \code{R6}) of length 19.
}
\usage{
tour_guide
}
\description{
Tour of the shiny app
}
\details{
\itemize{
\item \href{https://conductor.etiennebacher.com}{Conductor}
\item \href{https://github.com/carlganz/rintrojs}{rintrojs}
\item \href{https://cicerone.john-coene.com/index.html}{cicerone}
}
}
\keyword{datasets}
