% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{The application server-side}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}

\item{mod_values}{a 'reactiveValues() list storing returned values from server modules,
containing the following elements
\itemize{
\item 'upload_data' uploaded csv file as data frame.
\item 'species_selected'' species select in selectInput input control.
\item 'species_year' year selected in numericInput input control.
}'filter_data' uploaded csv file filtered by species and year list.}
}
\description{
The application server-side
}
