% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_display_plot.R
\name{mod_display_plot_server}
\alias{mod_display_plot_server}
\title{Shiny module display plot server function}
\usage{
mod_display_plot_server(id, mod_values)
}
\description{
This shiny module creates and displays a plot of the filtered data on top of the upload data,
\itemize{
\item check validated filter data avaible and has permission to display plot.
\item creates a plot of the filtered data on top of the upload data.
\item displays the plot of the filtered data on top of the upload data.
\item provides brush option displaying the selected data in a table.
}
}
\details{
\itemize{
\item \href{https://plotly-r.com/}{plotly}
\item \href{https://gallery.shinyapps.io/095-plot-interaction-advanced}{plot input}
\itemize{
\item nearPoints: plot_click, plot_dblclick, plot_hover
\item brushedPoints: plot_brush
}
}
}
