% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_display_plot.R
\name{mod_display_plot_server}
\alias{mod_display_plot_server}
\title{Shiny module display plot server function}
\usage{
mod_display_plot_server(id, mod_values)
}
\description{
This shiny module creates, displays and saves a plot of the filtered data on top of the upload data,
\itemize{
\item check validated filter data available and has permission to display plot.
\item check for presence of example invalid values for plot.
\item creates a plot of the filtered data on top of the upload data.
\item displays the plot of the filtered data on top of the upload data.
\item provides brush option displaying the selected data in a table.
\item adds a save button to save the plot.
}
}
\details{
\itemize{
\item \href{https://shiny.rstudio.com/reference/shiny/0.14/validate.html}{Validate input values}
\item \href{https://plotly-r.com/}{plotly}
\item \href{https://gallery.shinyapps.io/095-plot-interaction-advanced}{plot input}
\itemize{
\item nearPoints: plot_click, plot_dblclick, plot_hover
\item brushedPoints: plot_brush
}
}
}
