% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload.R
\name{mod_upload_server}
\alias{mod_upload_server}
\title{Shiny module upload server function}
\usage{
mod_upload_server(id, mod_values)
}
\arguments{
\item{id}{Internal parameter for {shiny}.}

\item{mod_values}{reactiveValues() list containing server module returned values}
}
\value{
upload_data() reactive value containing uploaded data from csv file.
}
\description{
Shiny module upload server function
}
\details{
\itemize{
\item \href{https://shiny.rstudio.com/reference/shiny/latest/reactivePoll.html}{Reactive polling}
\item \href{https://shiny.rstudio.com/reference/shiny/latest/reactiveFileReader.html}{Reactive file reader}
}
}
